requirements.txt
streamlit==1.20.0
google-auth==2.25.0
google-auth-oauthlib==0.5.2
google-auth-httplib2==0.1.0
google-api-python-client==2.88.0
google-api-core==2.11.1
python-dotenv==1.0.0
praw==7.7.0
requests==2.28.2
